refactor(execution-engine): extract task-runner-prepare from task-runner
Moves WorkRequest construction (buildWorkRequest: upstream gathering, dependency-branch guard, lifecycle tag, early branch persistence) out of task-runner into task-runner-prepare, behavior unchanged. executeTaskInner now delegates via buildWorkRequest(this, ...); the phase-host surface is widened/exported so the runner satisfies TaskRunnerPhaseHost.
One atomic slice of the step-4 large-file decomposition.
